+++ simplebluez/include/simplebluez/Characteristic.h
@@ -30,6 +30,7 @@ class Characteristic : public SimpleDBus::Proxy {
     std::string uuid();
     ByteArray value();
     bool notifying();
+    std::vector<std::string> flags();
 
     // ----- CALLBACKS -----
     void set_on_value_changed(std::function<void(ByteArray new_value)> callback);